3.02 - Restructure. Fix deployment of NPM module.
