+++ src/db/BasicObject.ts
@@ -1,0 +1,3 @@
+export interface BasicObject {
+  [key: string]: any;
+}
