+++ src/db/S3DBError.ts
@@ -1,0 +1,6 @@
+
+/**
+ * All exceptions thrown by this library are reported under this type.
+ */
+export class S3DBError extends Error {
+}
