+++ src/defaults/UUIDGenerator.ts
@@ -1,0 +1,9 @@
+import * as uuid4 from 'uuid/v4';
+import { BasicObject } from '../db';
+
+/**
+ * UUID v4.
+ */
+export function defaultIDGenerator(object: BasicObject): string {
+  return uuid4();
+}
