+++ src/defaults/index.ts
@@ -1,0 +1,3 @@
+export * from './JSONSerializer';
+export * from './MD5IsModified';
+export * from './UUIDGenerator';
