+++ src/exposed/IDGenerator.ts
@@ -1,0 +1,11 @@
+import { BasicObject } from "../db";
+
+/**
+ * Interface to be implemented to specify the generator pattern for
+ * a collection.
+ */
+export interface IDGenerator {
+  (object: BasicObject): string;
+}
+
+
