+++ src/exposed/IsModified.ts
@@ -1,0 +1,9 @@
+import { BasicObject } from "../db";
+
+export interface IsModified {
+  /**
+   * 
+   * @param documentToCheck if the difference is different from the head metadata.
+   */
+  isModified(documentToCheck: BasicObject, body: string): boolean;
+}
