+++ src/exposed/Serializers.ts
@@ -1,0 +1,7 @@
+/**
+ * Interface for serializing an object.
+ */
+export interface Serialization {
+  serialize(objectToSerialize: any): string;
+  deserialize<T>(documentToDeserialize: string): T;
+}
