+++ src/index.ts
@@ -1,8 +1,4 @@
-export { S3DB, S3DBConfiguration, S3DBError } from "./model/S3DB";
-export { id, Reference, ReferenceList, CollectionConfiguration, Collection, collection } from "./model/Collection";
-export { JSONSerialization, Serialization } from './model/Serializers'
-export { Validation } from './model/Validation';
-export { IDGenerator, defaultIDGenerator} from './model/IDGenerator';
-export { IsModified, MD5IsModified } from './model/IsModified';
-
-
+export * from './exposed';
+export * from './defaults';
+export * from './db';
+export * from './collection';
