+++ src/s3/index.ts
@@ -1,0 +1,3 @@
+export * from './Client';
+export * from './S3Metadata';
+export * from './S3Object';
